requirements.txt
pycryptodome
Pillow
pwntools
